requirements.txt
fastapi
uvicorn
pydantic
torch==2.1.2 
torchaudio==2.1.2 
torchvision==0.16.2
TTS
numpy
opencv-python
ffmpeg-python
transformers==4.36.2 
tokenizers==0.15.2
scipy
imageio
imageio-ffmpeg
librosa
tqdm
face-alignment
scikit-image
praw
feedparser
faiss-cpu 
sentence-transformers
python-dotenv
praw